+++ uarray/uarray/ast_types.py
@@ -1,0 +1,9 @@
+import ast
+
+
+class CStatement:
+    name: ast.AST
+
+
+class CIdentifier:
+    name: str
